+++ noether/_tokenizers.py
@@ -1,6 +1,7 @@
 from functools import wraps
+from collections import deque
 from io import BytesIO
-from typing import Any, Callable, Generator, Iterable, Iterator, Mapping
+from typing import Callable, Iterator
 from tokenize import tokenize, TokenInfo, NUMBER, NAME, OP, ENCODING
 
 
@@ -31,27 +32,33 @@ def transformer(
 @transformer
 def units_dialect(stream: TokenStream) -> TokenStream:
     '''
-    Process tokens for __main__ dialect, replacing `3u` -> `u(3)` and `in` -> `inch`.
+    Process tokens for __main__ dialect, replacing:
+    - `in` -> `inch`
+    - `-3unit` -> `unit(-3)`
     '''
-    num: TokenInfo | None = None
+    queue: deque[TokenInfo] = deque()
+
     for token in stream:
         if token.type == NAME and token.string == 'in':
             token = token._replace(string='inch')
 
-        if token.type == NUMBER:
-            num = token
-        elif num is not None:
-            if token.type == NAME:
-                yield token
+        queue.append(token)
+        if len(queue) == 3:
+            tt = [t.type for t in queue]
+            if tt[-2:] == [NUMBER, NAME]:
+                maybe_minus, number, unit_name = queue
+                is_minus = maybe_minus.type == OP and maybe_minus.string == '-'
+                if not is_minus:
+                    yield maybe_minus
+                yield unit_name
                 yield _t(OP, '(')
-                yield num
+                if is_minus:
+                    yield maybe_minus
+                yield number
                 yield _t(OP, ')')
+                queue.clear()
             else:
-                yield num
-                yield token
-            num = None
-        else:
-            yield token
+                yield queue.popleft()
 
 
 if __name__ == '__main__':