+++ noether/core/__init__.py
@@ -1,6 +1,7 @@
 from .Dimension import Dimension
 from .Prefix import Prefix
 from .MeasureInfo import MeasureInfo
+from .MeasureRelative import MeasureRelative
 from .Measure import Measure
 
 from .Unit import Unit