+++ noether/core/units/AffineUnit.py
@@ -15,12 +15,13 @@ class AffineUnit(Unit):
         zero_point: Measure,
         names: list[str] | str | None = None,
         symbols: list[str] | str | None = None,
+        info: str | None = None,
     ):
-        if unit.dim != zero_point.dim:
-            raise DimensionError(zero_point.dim, unit.dim,
-                                 '(AffineUnit creation)')
+        DimensionError.check(zero_point.dim, unit.dim, '(AffineUnit creation)')
         object.__setattr__(self, 'zero_point', zero_point)
-        Unit.__init__(self, unit, names, symbols)
+        Unit.__init__(
+            self, unit,
+            names=names, symbols=symbols, info=info)
 
     def __call__(self, value: Real, stddev: Real | None = None):
         return Measure.__call__(self, value, stddev) + self.zero_point
@@ -37,4 +38,4 @@ class AffineUnit(Unit):
         return Unit._repr_measure(self, measure - self.zero_point)
 
     def _json_extras(self):
-        return {'zero_point': self.zero_point / self._value}
+        return {'zero_point': self.zero_point._value}