+++ CHANGELOG.md
@@ -1,8 +1,28 @@
 # Noether changelog
 
 ## 1.1
+***28 Jul 2023, [44 commits](https://github.com/yunruse/Noether/pull/73/commits), closing [17 issues](https://github.com/yunruse/Noether/milestone/1?closed=1)***
 
-- Updated catalogue system
+- Various further bugfixes and improvements for display mechanisms
+- Support for logarithmic units, such as `decibel`
+  - Like with temperature `AffineUnit`s, call `dB(3)` or (`3dB` on the command-line) to use the scale; multiplication works as usual
+  - Decibel-units are also supported, e.g. `horsepower @ dB(kW)`
+- Time-based measures and units now interoperable with `datetime.timedelta`
+- Added 58 new units:
+  - **Cooking**: `tablespoon`, `teaspoon`, `dessertspoon`
+  - **Astronomy**: `solar_mass`, `solar_radius`, `solar_irradiance`, `solar_luminosity`, `solar_temperature`, `solar_mass_parameter`, `earth_radius_equatorial`, `earth_radius_polar`, `jupiter_radius_equatorial`, `jupiter_radius_polar`, `earth_mass_parameter`, `jupiter_mass_parameter`
+  - **Scientific**: `amagat`, `dalton`, `torr`, `bubnoff`, `eotvos`, `dBZ`
+  - **Historical scientific**:
+    - temperature: `degN`, `degRe`, `degDe`, `degRo`,
+    - radioactive: `curie`, `rutherford`, `thomson`, `roentgen`, `Mache`
+    - misc: `spat_length`, `foot_lambert`
+  - **Conventional**:
+    - logarithmic: `bel`, `decibel`, `neper`
+    - volume: `gill`, `quart`
+    - mass: `scruple`, `apdram`, `apounce`, `appound`, `pennyweight`
+    - mechanical: `mach`, `lbf`, `slug`, `psi`, `foot_pound`, `pound_foot`, `kip`
+    - misc: `msw`, `fsw`, `mired`, `julian_month`, `rpm`, `rps`, `mpg`
+  - **Unusual**: `flight_dose_rate`, `dog_year`
 
 
 ## 1.0.1
@@ -11,9 +31,8 @@
 - Various bugfixes and improved testing for display mechanisms; units now display in scientific notation where otherwise would be unwieldy.
 - Fixed incorrect value for `standard_gravity`.
 - Added a very rudimentary GNU `units`-like interface (`python -m noether.gnu`)
-- New units:
-  - `horsepower`
-  - `poundal`
+- Added 2 new units:
+  - `horsepower`, `poundal`
 - Deprecate unused internal features:
   - `MultiplicationWithValue`
   - "Ranged overlap" comparisons